backend/requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.4.0
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-ninja==1.1.0
dnspython==2.6.1
email_validator==2.2.0
gunicorn==22.0.0
idna==3.8
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.1.0
packaging==24.1
pillow==10.3.0
psycopg==3.2.1
psycopg-binary==3.2.1
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
rich==13.7.1
setuptools==68.2.2
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
wheel==0.41.2
